feat(react): add remote rspack module federation support
