+++ packages/react/src/generators/remote/files/rspack-module-federation-ssr/module-federation.server.config.js__tmpl__
@@ -1,0 +1,6 @@
+module.exports = {
+    name: '<%= projectName %>',
+   exposes: {
+     './Module': '<%= appProjectRoot %>/src/remote-entry.<%= js ? 'js' : 'ts' %>',
+   },
+};
