+++ packages/react/src/generators/remote/files/rspack-module-federation-ts/module-federation.config.ts__tmpl__
@@ -1,0 +1,13 @@
+import {ModuleFederationConfig} from '@nx/rspack';
+
+const config: ModuleFederationConfig = {
+    name: '<%= projectName %>',
+    <% if (dynamic) { %>
+   library: { type: 'var', name: '<%= projectName %>'},
+   <% } %>
+    exposes: {
+        './Module': './src/remote-entry.<%= js ? 'js' : 'ts' %>',
+    },
+};
+
+export default config;
