+++ packages/react/src/generators/remote/files/rspack-module-federation-ts/rspack.config.prod.ts__tmpl__
@@ -1,0 +1,1 @@
+export default require('./rspack.config');
