+++ packages/react/src/generators/remote/files/rspack-module-federation/module-federation.config.js__tmpl__
@@ -1,0 +1,9 @@
+module.exports = {
+   name: '<%= projectName %>',
+   <% if (dynamic) { %>
+   library: { type: 'var', name: '<%= projectName %>'},
+   <% } %>
+   exposes: {
+     './Module': './src/remote-entry.<%= js ? 'js' : 'ts' %>',
+   },
+};
