+++ packages/react/src/generators/remote/files/rspack-module-federation/rspack.config.prod.js__tmpl__
@@ -1,0 +1,1 @@
+module.exports = require('./rspack.config');
