+++ packages/react/src/generators/remote/files/webpack-module-federation-ssr-ts/server.ts__tmpl__
