+++ packages/react/src/generators/remote/files/webpack-module-federation-ssr/module-federation.server.config.js__tmpl__
