+++ packages/react/src/generators/remote/files/webpack-module-federation-ssr/server.ts__tmpl__
