+++ packages/react/src/generators/remote/files/webpack-module-federation-ssr/webpack.server.config.js__tmpl__
