+++ packages/react/src/generators/remote/files/webpack-module-federation-ts/module-federation.config.ts__tmpl__
