+++ packages/react/src/generators/remote/files/webpack-module-federation-ts/tsconfig.lint.json__tmpl__
