+++ packages/react/src/generators/remote/files/webpack-module-federation-ts/webpack.config.prod.ts__tmpl__
