+++ packages/react/src/generators/remote/files/webpack-module-federation-ts/webpack.config.ts__tmpl__
