+++ packages/react/src/generators/remote/files/webpack-module-federation/module-federation.config.js__tmpl__
