+++ packages/react/src/generators/remote/files/webpack-module-federation/webpack.config.js__tmpl__
