+++ packages/react/src/generators/remote/files/webpack-module-federation/webpack.config.prod.js__tmpl__
