+++ packages/react/src/rules/update-module-federation-project.ts
@@ -18,42 +18,82 @@ export function updateModuleFederationProject(
     devServerPort?: number;
     typescriptConfiguration?: boolean;
     dynamic?: boolean;
+    bundler?: 'rspack' | 'webpack';
   }
 ): GeneratorCallback {
   const projectConfig = readProjectConfiguration(host, options.projectName);
 
-  projectConfig.targets.build.options = {
-    ...projectConfig.targets.build.options,
-    main: maybeJs(options, `${options.appProjectRoot}/src/main.ts`),
-    webpackConfig: `${options.appProjectRoot}/webpack.config.${
-      options.typescriptConfiguration && !options.js ? 'ts' : 'js'
-    }`,
-  };
+  if (options.bundler === 'rspack') {
+    projectConfig.targets.build.executor = '@nx/rspack:rspack';
+    projectConfig.targets.build.options = {
+      ...projectConfig.targets.build.options,
+      main: maybeJs(options, `${options.appProjectRoot}/src/main.ts`),
+      rspackConfig: `${options.appProjectRoot}/rspack.config.${
+        options.typescriptConfiguration && !options.js ? 'ts' : 'js'
+      }`,
+    };
 
-  projectConfig.targets.build.configurations.production = {
-    ...projectConfig.targets.build.configurations.production,
-    webpackConfig: `${options.appProjectRoot}/webpack.config.prod.${
-      options.typescriptConfiguration && !options.js ? 'ts' : 'js'
-    }`,
-  };
+    projectConfig.targets.build.configurations.production = {
+      ...projectConfig.targets.build.configurations.production,
+      rspackConfig: `${options.appProjectRoot}/rspack.config.prod.${
+        options.typescriptConfiguration && !options.js ? 'ts' : 'js'
+      }`,
+    };
+  } else {
+    projectConfig.targets.build.options = {
+      ...projectConfig.targets.build.options,
+      main: maybeJs(options, `${options.appProjectRoot}/src/main.ts`),
+      webpackConfig: `${options.appProjectRoot}/webpack.config.${
+        options.typescriptConfiguration && !options.js ? 'ts' : 'js'
+      }`,
+    };
+
+    projectConfig.targets.build.configurations.production = {
+      ...projectConfig.targets.build.configurations.production,
+      webpackConfig: `${options.appProjectRoot}/webpack.config.prod.${
+        options.typescriptConfiguration && !options.js ? 'ts' : 'js'
+      }`,
+    };
+  }
 
   // If host should be configured to use dynamic federation
   if (options.dynamic) {
-    const pathToProdWebpackConfig = joinPathFragments(
-      projectConfig.root,
-      `webpack.prod.config.${
-        options.typescriptConfiguration && !options.js ? 'ts' : 'js'
-      }`
-    );
-    if (host.exists(pathToProdWebpackConfig)) {
-      host.delete(pathToProdWebpackConfig);
-    }
+    if (options.bundler === 'rspack') {
+      const pathToProdRspackConfig = joinPathFragments(
+        projectConfig.root,
+        `rspack.prod.config.${
+          options.typescriptConfiguration && !options.js ? 'ts' : 'js'
+        }`
+      );
+      if (host.exists(pathToProdRspackConfig)) {
+        host.delete(pathToProdRspackConfig);
+      }
 
-    delete projectConfig.targets.build.configurations.production?.webpackConfig;
+      delete projectConfig.targets.build.configurations.production
+        ?.rspackConfig;
+    } else {
+      const pathToProdWebpackConfig = joinPathFragments(
+        projectConfig.root,
+        `webpack.prod.config.${
+          options.typescriptConfiguration && !options.js ? 'ts' : 'js'
+        }`
+      );
+      if (host.exists(pathToProdWebpackConfig)) {
+        host.delete(pathToProdWebpackConfig);
+      }
+
+      delete projectConfig.targets.build.configurations.production
+        ?.webpackConfig;
+    }
   }
 
-  projectConfig.targets.serve.executor =
-    '@nx/react:module-federation-dev-server';
+  if (options.bundler === 'rspack') {
+    projectConfig.targets.serve.executor =
+      '@nx/rspack:module-federation-dev-server';
+  } else {
+    projectConfig.targets.serve.executor =
+      '@nx/react:module-federation-dev-server';
+  }
   projectConfig.targets.serve.options.port = options.devServerPort;
 
   // `serve-static` for remotes that don't need to be in development mode
